Allocate Program UID for MIC internally without BE.

+++ opencl-intel/devices/mic_device/mic_native/native_program_service.h
@@ -149,7 +149,7 @@ private:
         TKernelList             kernels;
     };
 
-    typedef std::map<unsigned long long int, TProgramEntry*> TProgId2Map;
+    typedef std::map<uint64_t, TProgramEntry*> TProgId2Map;
 
     BackendInterfaces              m_BE_Executor;
